+++ cpu_driver.h
@@ -40,7 +40,7 @@ struct cpu_driver_funcs {
 
   void (*memory_range_invalidate)(struct cpu_driver* p_cpu_driver,
                                   uint16_t addr,
-                                  uint16_t len);
+                                  uint32_t len);
   char* (*get_address_info)(struct cpu_driver* p_cpu_driver, uint16_t addr);
   void (*get_custom_counters)(struct cpu_driver* p_cpu_driver,
                               uint64_t* p_c1,
